+++ RELEASE-NOTES.md
@@ -2,6 +2,9 @@
 ### Version 5.0.25-case-file-view-error-message-display
 **EUI-6826** Display standard error message if Case File View component is unable to retrieve categories and documents data
 
+### Version 5.0.25-case-file-view-media-viewer-integration
+**EUI-5349** Integrate the Media Viewer component with Case File View
+
 ### Version 5.0.24-case-file-view-folder-document-icons-v3
 **EUI-6974** Case file view folder and document icons display